remove some warnings and duplicated code
