+++ src/TortoiseShell/TortoiseStub.vcxproj
@@ -42,7 +42,7 @@
   </PropertyGroup>
   <ItemDefinitionGroup>
     <ClCompile>
-      <AdditionalIncludeDirectories>..\Utils;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
+      <AdditionalIncludeDirectories>$(ProjectDir);..\Utils;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
       <PreprocessorDefinitions>TORTOISESTUB_EXPORTS;%(PreprocessorDefinitions)</PreprocessorDefinitions>
       <PrecompiledHeader>NotUsing</PrecompiledHeader>
     </ClCompile>
@@ -62,6 +62,7 @@
     </ClCompile>
   </ItemDefinitionGroup>
   <ItemGroup>
+    <ClCompile Include="..\Utils\EventLog.cpp" />
     <ClCompile Include="TortoiseStub.cpp" />
   </ItemGroup>
   <ItemGroup>
@@ -69,6 +70,10 @@
   </ItemGroup>
   <ItemGroup>
     <ResourceCompile Include="TortoiseStub.rc" />
+  </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="..\Utils\EventLog.h" />
+    <ClInclude Include="stdafx.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
