+++ src/TortoiseShell/TortoiseStub.vcxproj.filters
@@ -13,10 +13,16 @@
       <UniqueIdentifier>{20d5bf6b-6b47-4967-a7c3-407e72e2acd0}</UniqueIdentifier>
       <Extensions>ico;cur;bmp;dlg;rc2;rct;bin;rgs;gif;jpg;jpeg;jpe</Extensions>
     </Filter>
+    <Filter Include="Utils">
+      <UniqueIdentifier>{9d533f37-aa5e-45f0-beb5-654b302e1cad}</UniqueIdentifier>
+    </Filter>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="TortoiseStub.cpp">
       <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="..\Utils\EventLog.cpp">
+      <Filter>Utils</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
@@ -28,5 +34,13 @@
     <ResourceCompile Include="TortoiseStub.rc">
       <Filter>Resource Files</Filter>
     </ResourceCompile>
+  </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="..\Utils\EventLog.h">
+      <Filter>Utils</Filter>
+    </ClInclude>
+    <ClInclude Include="stdafx.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
 </Project>
