+++ src/Utils/EventLog.cpp
@@ -16,10 +16,13 @@
 // along with this program; if not, write to the Free Software Foundation,
 // 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 //
+
 #include "stdafx.h"
+#include <Windows.h>
 #include "EventLog.h"
 
 namespace EventLog {
+	static const wchar_t* APPLICATION_NAME = L"TortoiseSI";
 
 	//
 	//   FUNCTION: EventLog::writeEvent(wchar_t *, WORD)
@@ -38,7 +41,7 @@ namespace EventLog {
 	//     EVENTLOG_INFORMATION_TYPE
 	//     EVENTLOG_WARNING_TYPE
 	//
-	void writeEvent(const wchar_t *message, WORD wType) {
+	void writeEvent(std::wstring message, WORD wType) {
 		HANDLE hEventSource = NULL;
 		LPCWSTR lpszStrings[2] = {NULL, NULL};
 		const wchar_t *name = EventLog::APPLICATION_NAME;
@@ -46,7 +49,7 @@ namespace EventLog {
 		hEventSource = RegisterEventSource(NULL, name);
 		if (hEventSource) {
 			lpszStrings[0] = name;
-			lpszStrings[1] = message;
+			lpszStrings[1] = message.c_str();
 
 			 ReportEvent(hEventSource,		// Event log handle
 				wType,						// Event type
@@ -72,7 +75,7 @@ namespace EventLog {
 	//   * error - the error message
 	//
 	void writeError(std::wstring error) {
-		writeEvent(error.c_str(), EVENTLOG_ERROR_TYPE);
+		writeEvent(error, EVENTLOG_ERROR_TYPE);
 	}
 	//
 	//   FUNCTION: EventLog::writeInformationLogEntry(wstring)
@@ -83,7 +86,7 @@ namespace EventLog {
 	//   * info - the information message
 	//
 	void writeInformation(std::wstring info) {
-		writeEvent(info.c_str(), EVENTLOG_INFORMATION_TYPE);
+		writeEvent(info, EVENTLOG_INFORMATION_TYPE);
 	}
 
 	//
@@ -95,6 +98,13 @@ namespace EventLog {
 	//   * warning - the warning message
 	//
 	void writeWarning(std::wstring warning) {
-		writeEvent(warning.c_str(), EVENTLOG_WARNING_TYPE);
+		writeEvent(warning, EVENTLOG_WARNING_TYPE);
 	}
+}
+
+std::wstring getProcessFilesName()
+{
+	wchar_t moduleName[MAX_PATH] = { 0 };
+	GetModuleFileName(NULL, moduleName, _countof(moduleName));
+	return std::wstring(moduleName);
 }
